Integrate notification into ModelService class
- Removed redundant ExampleService and ExampleApp
