--- app/src/main/java/com/example/unblind/ExampleApp.java
@@ -1,30 +1,0 @@
-package com.example.unblind;
-
-import android.app.Application;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
-import android.os.Build;
-
-public class ExampleApp extends Application {
-    public static final String CHANNEL_ID = "exampleChannel";
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        createNotificationChannel();
-    }
-
-    private void createNotificationChannel() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            CharSequence name = getString(R.string.channel_name);
-            String description = getString(R.string.channel_description);
-            int importance = NotificationManager.IMPORTANCE_DEFAULT;
-
-            NotificationChannel exampleChannel = new NotificationChannel(CHANNEL_ID, name, importance);
-            exampleChannel.setDescription(description);
-
-            NotificationManager notificationManager = getSystemService(NotificationManager.class);
-            notificationManager.createNotificationChannel(exampleChannel);
-        }
-    }
-}
