--- app/src/main/java/com/example/unblind/ExampleService.java
@@ -1,88 +1,0 @@
-package com.example.unblind;
-
-import android.app.Notification;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
-import android.app.Service;
-import android.content.Context;
-import android.content.Intent;
-import android.os.Binder;
-import android.os.Build;
-import android.os.IBinder;
-import android.util.Log;
-import android.widget.Toast;
-
-import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
-import androidx.core.app.NotificationCompat;
-
-public class ExampleService extends Service {
-    private int NOTIFICATION = R.string.example_service_running;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        startNotification();
-    }
-
-    @Override
-    public int onStartCommand(Intent intent, int flags, int startId) {
-        super.onStartCommand(intent, flags, startId);
-        if(intent.getAction() != null && intent.getAction().equals(getString(R.string.turn_off))) {
-            stopForeground(true);
-        }
-        return START_NOT_STICKY;
-    }
-
-    @Override
-    public void onDestroy() {
-        super.onDestroy();
-    }
-
-    @Nullable
-    @Override
-    public IBinder onBind(Intent intent) {
-        return null;
-//        return exampleBinder;
-    }
-
-    private void startNotification() {
-        CharSequence text = getText(R.string.example_service_running);
-
-        PendingIntent contentIntent = PendingIntent.getActivity(this, 0,
-                new Intent(this, MainActivity.class), 0);
-        Intent stopSelf = new Intent(this, ExampleService.class);
-        stopSelf.setAction(getString(R.string.turn_off));
-        PendingIntent pStopSelf = PendingIntent.getService(this, 0, stopSelf,PendingIntent.FLAG_CANCEL_CURRENT);
-
-        NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-        String channelId = Build.VERSION.SDK_INT >= Build.VERSION_CODES.O ? createNotificationChannel(notificationManager) : "";
-        NotificationCompat.Builder notificationBuilder = new NotificationCompat.Builder(this, channelId);
-        Notification notification = notificationBuilder.setOngoing(true)
-                .setSmallIcon(R.drawable.ic_android_black_24dp)
-                .setTicker(text)
-                .setWhen(System.currentTimeMillis())
-                .setContentTitle(getText(R.string.example_service_label))
-                .setContentText(text)
-                .setContentIntent(contentIntent)
-                .setCategory(NotificationCompat.CATEGORY_SERVICE)
-                .addAction(R.drawable.ic_android_black_24dp, getString(R.string.turn_off), pStopSelf)
-                .build();
-
-        startForeground(1, notification);
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private String createNotificationChannel(NotificationManager notificationManager){
-        String channelId = "exampleChannel";
-        String channelName = "Example Channel";
-        NotificationChannel channel = new NotificationChannel(channelId, channelName, NotificationManager.IMPORTANCE_HIGH);
-        // omitted the LED color
-        channel.setImportance(NotificationManager.IMPORTANCE_NONE);
-        channel.setLockscreenVisibility(Notification.VISIBILITY_PRIVATE);
-        notificationManager.createNotificationChannel(channel);
-        return channelId;
-    }
-
-}
