+++ app/src/main/java/com/example/unblind/ModelService.java
@@ -1,5 +1,9 @@
 package com.example.unblind;
 
+import android.app.Notification;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.app.PendingIntent;
 import android.app.Service;
 import android.content.ComponentName;
 import android.content.Context;
@@ -7,11 +11,14 @@ import android.content.Intent;
 import android.content.ServiceConnection;
 import android.graphics.Bitmap;
 import android.os.AsyncTask;
+import android.os.Build;
 import android.os.IBinder;
 import android.util.Log;
 import android.util.Pair;
 
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
+import androidx.core.app.NotificationCompat;
 import androidx.work.OneTimeWorkRequest;
 import androidx.work.WorkManager;
 
@@ -36,8 +43,7 @@ public class ModelService extends Service implements ColleagueInterface {
             String absolutePath = Utils.assetFilePath(getOuter(), strings[0]); //get absolute path
             return new Classifier(absolutePath);
         }
-
-        @Override
+            @Override
         protected void onPostExecute(Classifier result) {
             Log.e(TAG, "onPostExecute: ");
             setClassifier(result);
@@ -92,6 +98,12 @@ public class ModelService extends Service implements ColleagueInterface {
     }
 
     @Override
+    public void onCreate() {
+        super.onCreate();
+        startNotification();
+    }
+
+    @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
         Log.e(TAG, "Model service started");
         super.onStartCommand(intent, flags, startId);
@@ -101,6 +113,9 @@ public class ModelService extends Service implements ColleagueInterface {
         Intent newIntent = new Intent(this, DatabaseService.class);
         startService(newIntent);
         bindService(newIntent, mConnection, Context.BIND_AUTO_CREATE);
+        if(intent.getAction() != null && intent.getAction().equals(getString(R.string.turn_off))) {
+            stopForeground(true);
+        }
         return START_NOT_STICKY;
     }
 
@@ -118,6 +133,44 @@ public class ModelService extends Service implements ColleagueInterface {
         Log.d("Team 3 Model Result", result);
         currentElement = new Pair<Bitmap, String>(currentElement.first, result);
         mediator.setElement(currentElement);
+    }
+
+    private void startNotification() {
+        CharSequence text = getText(R.string.example_service_running);
+
+        PendingIntent contentIntent = PendingIntent.getActivity(this, 0,
+                new Intent(this, MainActivity.class), 0);
+        Intent stopSelf = new Intent(this, ModelService.class);
+        stopSelf.setAction(getString(R.string.turn_off));
+        PendingIntent pStopSelf = PendingIntent.getService(this, 0, stopSelf,PendingIntent.FLAG_CANCEL_CURRENT);
+
+        NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        String channelId = Build.VERSION.SDK_INT >= Build.VERSION_CODES.O ? createNotificationChannel(notificationManager) : "";
+        NotificationCompat.Builder notificationBuilder = new NotificationCompat.Builder(this, channelId);
+        Notification notification = notificationBuilder.setOngoing(true)
+                .setSmallIcon(R.drawable.ic_android_black_24dp)
+                .setTicker(text)
+                .setWhen(System.currentTimeMillis())
+                .setContentTitle(getText(R.string.example_service_label))
+                .setContentText(text)
+                .setContentIntent(contentIntent)
+                .setCategory(NotificationCompat.CATEGORY_SERVICE)
+                .addAction(R.drawable.ic_android_black_24dp, getString(R.string.turn_off), pStopSelf)
+                .build();
+
+        startForeground(1, notification);
+    }
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    private String createNotificationChannel(NotificationManager notificationManager){
+        String channelId = "exampleChannel";
+        String channelName = "Example Channel";
+        NotificationChannel channel = new NotificationChannel(channelId, channelName, NotificationManager.IMPORTANCE_HIGH);
+        // omitted the LED color
+        channel.setImportance(NotificationManager.IMPORTANCE_NONE);
+        channel.setLockscreenVisibility(Notification.VISIBILITY_PRIVATE);
+        notificationManager.createNotificationChannel(channel);
+        return channelId;
     }
 
 }
